ml/requirements.txt
tensorflow>=2.13.0
# For GPU support (NVIDIA GPUs like RTX 5070):
# tensorflow[and-cuda]>=2.13.0  # Includes CUDA dependencies
keras>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
